+++ Backend/Utilities/Verification.py
@@ -3,6 +3,7 @@ import logging
 import os
 
 import jwt
+from deprecated.classic import deprecated
 
 from sendgrid import SendGridAPIClient
 from sendgrid.helpers.mail import Mail
@@ -11,14 +12,15 @@ from Constants.Constants import SENDGRID_API_KEY
 from Data.Neo4j.NodeDatabaseManagement import NodeDatabaseManagement as NodeDB
 
 
+@deprecated
 def send_verification_email(email: str):
     """
+    #ToDo: Password reset method
     Sends a verification email to the registered email address
 
     :param email: The entered email, (already passed check to confirm email isn't in database)
     """
     expiration_date = datetime.datetime.utcnow() + datetime.timedelta(hours=24)
-
 
     token = jwt.encode({
         'email': email,
@@ -30,12 +32,15 @@ def send_verification_email(email: str):
     message = Mail(
         from_email='verification@em6129.thethinkerai.com',
         to_emails=email,
-        subject='Verify your email for The Thinker AI',
+        subject='Verify your email for The Thinker AI - Get free credit',
         html_content=f'''
-            <p>Thank you for registering with The Thinker!</p>
-            <p>Please verify your email by clicking the link below:</p>
+            <p>Thank you for registering with The Thinker AI!</p>
+            <p>Please verify you own this email by clicking below</p>
             <a href="{verification_link}">Verify Email</a>
             <small>This link will expire in 24 hours.</small>
+            
+            <br>
+            <p>When verified, we will give you a free dollar of credit to get started.</p>
             
             <p>
                 <br>
@@ -69,7 +74,12 @@ def apply_new_user_promotion(email: str):
 
         if remaining_new_user_promotions is not None:
             promotion_applied = True
+            logging.info(f"Promotion applied for {email}")
             if remaining_new_user_promotions == 0:
                 last_promotion = True
+        else:
+            logging.info("1$ Promotion could not be applied")
+    else:
+        logging.error("New user not found in database!")
 
     return promotion_applied, last_promotion